move resolve_json_refs to prompt_utils to avoid importing fastapi in training
