+++ functionary/inference_utils.py
@@ -1,8 +1,6 @@
-from copy import deepcopy
 from http import HTTPStatus
 from typing import Dict, List, Optional
 
-import jsonref
 import torch
 from fastapi.responses import JSONResponse
 from pydantic import BaseModel
@@ -112,23 +110,6 @@ async def check_all_errors(request, served_model) -> Optional[JSONResponse]:
             param="tool_choice",
         )
     return
-
-
-def resolve_json_refs(tools_or_functions):
-    tools = deepcopy(tools_or_functions)
-    if tools:
-        for i in range(len(tools)):
-            if "type" in tools[i]:
-                if tools[i]["type"] == "function":
-                    tools[i]["function"]["parameters"] = deepcopy(
-                        jsonref.JsonRef.replace_refs(tools[i]["function"]["parameters"])
-                    )
-            else:
-                tools[i]["parameters"] = deepcopy(
-                    jsonref.JsonRef.replace_refs(tools[i]["parameters"])
-                )
-
-    return tools
 
 
 def convert_tool_calls_to_function_call(
