+++ functionary/vllm_monkey_patch/async_llm_engine.py
@@ -56,7 +56,7 @@ from vllm.utils import deprecate_kwargs, weak_bind
 from functionary.inference import (
     get_lm_format_enforcer_vllm_logits_processor_from_tool_name,
 )
-from functionary.inference_utils import resolve_json_refs
+from functionary.prompt_template.prompt_utils import resolve_json_refs
 from functionary.openai_types import Tool
 
 logger = init_logger(__name__)
